Move test from Bugs -> DicomClientTests
